feat: доработка стилей содержимого страницы
