+++ src/i18n/langs/ru.json
@@ -72,6 +72,7 @@
     "pluto": "Плутон"
   },
   "detail": {
+    "onlyFutureDates": "Только будущие даты",
     "orbitingBody": "Тело орбиты",
     "orbitalData": {
       "word": "Данные орбиты",
