+++ src/pages/Detail/common.ts
@@ -1,0 +1,1 @@
+export const APPROACH_DATA_LIMIT = 30;
